WIP on stop bit system. Core seems to work again when bundleAlignLog2 = 3.
Got rid of the idea of having a branch unit in every lane which can support
a stop bit. Instead, there's still always only one branch unit per lane
group, which must be in the last physical lane. When a stop bit is detected
and the associated syllable is a branch instruction, the syllable is sent to
the last lane, along with the PC+1 value (which is still computed per lane).
This is possible because there are no branch instructions with long
immediates (otherwise the LIMMH forwarding would need to be modified). I
think this will result in less hardware as well as make it faster, because
the PC+1 routing is less complicated (it always needs to route to the last
lane, whereas it had to route to all lanes before) and the extra muxing
between branch unit outputs is avoided because there's always only one.