Added all-at-once program, which combines all benchmarks into a single program.
Instead of having separate c files for each benchmark in which the main() is
renamed, objcopy trickery is used to add prefixes to all global symbols defined
by the subprograms, and a python script generates an assembly file containing
the actual main. It should work, but the standalone system does not have enough
memory, so I haven't been able to test it yet.