Partially fixed data cache control unit
There seems to be a (performance) issue with the data cache, where it sometimes
misses when it shouldn't, or handles a write in some block while another block
already has the value in storage (which is not allowed). The fix here fixes at
least one instance of the problem, where the write priority was not being
computed properly.