Fix incorrect pointer arithmetic on 64-bit systems.
In the expression `buffer + (j.base - address)`, if `j.base < address`
the result of `j.base - address` will be a large 32-bit value. If the
system has 64-bit pointers, the result of the original expression is
equal to `buffer + j.base - address + 1 << 32`, which is not intended.

This is fixed by changing the expression to `buffer + j.base - address`,
where the subtraction will be performed on the 64-bit value.