Disabled data cache flushing feature because of synthesis
Without the flush input, this component will be synthesized to a BRAM.
With the flush input, XST will do an optimization that takes forever (and is probably exponential in complexity with the cache size).
It would be nice to synthesize this to registers without the stupid optimization,
because that would still allow flushing (RAMBs are not cleared at reset).