Make the merged pcie branch with master branch synthesize again.
Mostly this involved removing part of the pcie register interface.
Everything was removed from the register interface except for
the ability to read the number of cores and contexts (which assumes each
core has the same number of contexts. This interface should probably be removed
entirely, or changed to better integrate with the base design.

Also fix the config.vhd with correct number of contexts for this platform, and
the correct value for the clock divider, since the pcie platform uses a
different clock source.

Revert "Fix run control reset vector functionality."
This reverts commit 5b1c3523ea53d283b7649866593d2521afb0e203.

Revert "Add register interface." (partially)
This reverts commit d1786aecc7ca5e8e70f295b405c596ae7f6b432d.
Only changes to the following files were reverted:
        lib/rvex/core/core.vhd
        lib/rvex/core/core_pkg.vhd
        lib/rvex/system/rvsys_grlib.vhd