Updated zedboard platform.
Linux should now work headless, i.e. without PL support. It should boot
automatically when the USB/UART cable is not connected (the default Linux
drivers seem to send garbage to the board for some reason, which aborts the
autoboot process). There is now also a make ssh target, which will attempt
to connect to the board (given a hostname). Unfortunately Linaro does not
ship with SSH enabled or even installed, so currently you still need to
install that manually using the UART terminal after imaging the card.