Ran into a (the?) cache bug on standalone-cached in a very reproducable way:
 - Synthesize the design with the CFG vector currently specified in ml605.vhd
 - make run-cache-test DYNAMIC=true-O1 BUNDLE_ALIGN=2 LANECONFIG=FBFBFBFB BUSLAT=20
 - qurt fails at high interrupt rates

This ONLY happens with the cached version of the standalone platform, and it
only seems to happen when reconfiguring without affecting the context qurt runs
in (for example 0x1188 to 0x1180, qurt in context 1). It also only seems to
happen at high interrupt rates, but that's probably just random chance.