Fixed zed-almarvi.
The VHDL design files for the r-VEX were actually missing from the place where
the platform was looking for them. They were never committed because they were
just copies of the r-VEX sources but with the library name changed to work, and
having two versions of the code in the repo would be bad. I forgot to make a
generate script for them though, which is now done. I also made the CFG
specification forwards compatible as it wasn't working anymore and I thus had
to fix it anyway.